test: Move setEnv & unsetEnv functions
